requirements.txt
tqdm>=4.66
python-dotenv>=1.0

google-generativeai>=0.8.5
